requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.2
aiosignal==1.3.1
alabaster==0.7.15
annotated-types==0.6.0
anyio==3.6.2
arabic-reshaper==3.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asn1crypto==1.5.1
astroid==2.15.8
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.1
autopep8==1.6.0
Babel==2.14.0
beautifulsoup4==4.13.4
black==22.12.0
blinker==1.9.0
certifi==2024.7.4
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.0.1
click==8.1.7
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.2.0
coverage==7.0.0
crispy-bootstrap5==0.7
cryptography==44.0.0
cssselect2==0.7.0
curl_cffi==0.11.4
cycler==0.12.1
Cython==3.0.7
decorator==5.1.1
defusedxml==0.8.0rc2
Deprecated==1.2.14
dill==0.3.7
distlib==0.3.8
distro==1.8.0
dj-database-url==2.1.0
Django==5.0.8
django-appconf==1.0.6
django-cleanup==6.0.0
django-compressor==4.1
django-cors-headers==3.13.0
django-coverage-plugin==3.0.0
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-filter==23.5
django-froala-editor==4.0.14
django-heroku==0.3.1
django-jet-reboot==1.3.5
django-model-utils==4.3.1
django-redis==5.2.0
django-stubs==1.13.1
django-stubs-ext==4.2.7
django-templated-mail==1.1.1
django-widget-tweaks==1.5.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==1.8.0
djoser==2.2.2
docutils==0.19
drf-spectacular==0.25.1
easyocr==1.7.0
executing==2.0.1
factory-boy==3.2.1
Faker==22.1.0
Farama-Notifications==0.0.4
filelock==3.13.1
filterpy==1.4.5
flake8==6.0.0
flake8-isort==5.0.3
Flask==3.1.0
flask-cors==5.0.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.48.1
fpdf==1.7.2
frozendict==2.4.6
frozenlist==1.3.3
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
gopay==2.0.1
greenlet==3.1.1
grpcio==1.64.1
gymnasium==1.2.3
h11==0.14.0
h2==4.1.0
h5py==3.11.0
hiredis==2.1.0
hpack==4.0.0
html5lib==1.1
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.29.3
hyperframe==6.0.1
identify==2.5.33
idna==3.7
imageio==2.34.2
imagesize==1.4.1
imutils==0.5.4
inflection==0.5.1
iniconfig==2.0.0
intel-openmp==2021.4.0
ipdb==0.13.11
ipython==8.20.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
keras==3.9.0
kiwisolver==1.4.5
lazy-object-proxy==1.10.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
limits==5.5.0
livereload==2.6.3
llvmlite==0.44.0
lxml==5.3.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mkl==2021.4.0
ml_dtypes==0.5.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.4
multitasking==0.0.11
mypy==0.982
mypy-extensions==1.0.0
namex==0.0.8
narwhals==2.2.0
networkx==3.3
ninja==1.11.1.1
nodeenv==1.8.0
numba==0.61.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.6.1
openai-whisper==20240930
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.11.0
ordered-set==4.1.0
oscrypto==1.3.0
package_name==0.1
packaging==23.2
panda==0.3.1
pandas==2.2.3
parso==0.8.3
pathspec==0.12.1
peewee==3.18.1
Pillow==9.4.0
platformdirs==4.1.0
plotly==6.3.0
pluggy==1.3.0
pooch==1.8.2
pre-commit==2.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==6.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.5
pure-eval==0.2.2
pyclipper==1.3.0.post5
pycodestyle==2.8.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyflakes==3.0.1
Pygments==2.17.2
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyJWT==2.8.0
pylint==2.17.7
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
PyMySQL==1.0.2
pyparsing==3.1.1
pypdf==5.2.0
pytesseract==0.3.10
pytest==7.2.0
pytest-django==4.5.2
pytest-sugar==0.9.6
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.21.1
python-slugify==7.0.0
python-telegram-bot==20.1
python3-openid==3.2.0
pytz==2022.7
PyYAML==6.0.1
qrcode==8.0
rcssmin==1.1.0
redis==4.4.0
referencing==0.32.1
regex==2024.11.6
reportlab==4.0.4
requests==2.32.3
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.7.1
rjsmin==1.2.0
rpds-py==0.16.2
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.10.1
seaborn==0.13.2
shapely==2.0.4
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
Sphinx==5.3.0
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
SQLAlchemy==2.0.40
sqlparse==0.5.0
stable_baselines3==2.7.1
stack-data==0.6.3
stripe==5.5.0
svglib==1.5.1
sympy==1.12.1
tbb==2021.13.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
text-unidecode==1.3
tf_keras==2.19.0
threadpoolctl==3.6.0
tifffile==2024.6.18
tiktoken==0.9.0
tinycss2==1.4.0
tk==0.1.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.3
torch==2.3.1
torchvision==0.18.1
tornado==6.4
tqdm==4.66.3
traitlets==5.14.1
transformers==4.49.0
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
types-requests==2.31.0.20240106
typing_extensions==4.9.0
tzdata==2025.2
tzlocal==5.2
ultralytics==8.0.114
uritemplate==4.1.1
uritools==4.0.3
urllib3==1.26.19
virtualenv==20.25.0
watchdog==3.0.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.2.0
wrapt==1.16.0
xgboost==3.0.4
xhtml2pdf==0.2.16
yarl==1.8.2
yfinance==0.2.63
